Moved update hook tests into hook_tests
